requirements.txt
anyio==4.6.2.post1
azure-core==1.26.4
azure-cosmos==4.3.1
backoff==2.2.1
bcrypt==4.0.1
blinker==1.9.0
cachelib==0.13.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
coloredlogs==15.0.1
exceptiongroup==1.2.2
filelock==3.16.1
Flask==2.3.2
Flask-Caching==2.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
flatbuffers==24.3.25
fsspec==2024.10.0
h11==0.14.0
httptools==0.6.4
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
monotonic==1.6
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
posthog==3.7.2
protobuf==5.28.3
pulsar-client==3.5.0
pydantic==1.10.19
PyPika==0.48.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.28.2
six==1.16.0
sniffio==1.3.1
starlette==0.27.0
sympy==1.13.1
threadpoolctl==3.5.0
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
uvicorn==0.32.1
watchfiles==0.24.0
websockets==14.1
Werkzeug==2.3.6
zipp==3.21.0
